requirements.txt
numpy
opencv-python
tensorflow
scikit-learn
matplotlib
pillow
Flask
Flask-RESTful
requests
PyJWT
python-dotenv
torch
torchvision
transformers
pytesseract
pdf2image
h5py
tqdm
imutils
efficientnet
gdown
keras-tuner
tensorboard
pytest 